Makes running of data.json the norm
This updates everything to run directly off data.json. It will immediately load to the home page by checking for a flag. that is now updated when creating the site.

Navigation is also fixed in this and the stucture of routing has been also altered.

Currently there is still a debug page which will be removed in the near future.